+++ jest.config.js
@@ -1,8 +1,4 @@
 module.exports = {
-  testEnvironment: 'node',
-  collectCoverage: true,
-  coverageDirectory: 'coverage',
-  coverageReporters: ['text', 'lcov'],
   coverageThreshold: {
     global: {
       branches: 80,
@@ -10,7 +6,5 @@ module.exports = {
       lines: 80,
       statements: 80
     }
-  },
-  testMatch: ['**/__tests__/**/*.js', '**/?(*.)+(spec|test).js'],
-  verbose: true
+  }
 };